+++ packages/pigeon/lib/src/objc/objc_generator.dart
@@ -2,11 +2,11 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'ast.dart';
-import 'functional.dart';
-import 'generator.dart';
-import 'generator_tools.dart';
-import 'pigeon_lib.dart' show Error, TaskQueueType;
+import '../ast.dart';
+import '../functional.dart';
+import '../generator.dart';
+import '../generator_tools.dart';
+import '../pigeon_lib.dart' show Error, TaskQueueType;
 
 /// Documentation comment open symbol.
 const String _docCommentPrefix = '///';